transportation_app/requirements.txt
flask
flask-socketio
eventlet
redis
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-firestore
google-cloud-storage
pandas
numpy
scikit-learn
joblib
argparse
